Pulled in files from Warlocks VR project. ProjectileData and Projectile. Neeed to implement an interface, tweak scripts, etc. All other code is working and this could be considered a safe revert point. #######

+++ Assets/Scripts/Units/Characters/ICharacter.cs
@@ -11,6 +11,6 @@ public interface ICharacter
     IStatusEffectHandler StatusEffectHandler { get; }
     GameObject GameObject { get; }
     IHealth Health { get; }
-    void Heal(int amount);
+    void Heal(float amount);
     event EventHandler<Character.HealedEventArgs> Healed;
 }
